Patch: Diff: Fix bug in parse_diff()
We do not understand changes of permissions yet. Commits of the linux
kernel like 06ed5c2bfacaf67039e87a213fa5d1cdde34246a broke the
evaluation as the Diff object contained 'empty hunks'.

This is only a quickfix. In future, we should also be able to parse
permission changes of files. For the sake of completeness.

Signed-off-by: Ralf Ramsauer <ralf.ramsauer@othr.de>