EquivalenceClass: Avoid collisions and allow to set None property
This is required when merging two equivalence classes that have
different properties, or when setting a property that already exists for
another class. In both cases, those classes will get merged.

So far, we forgot some housekeeping of the property_lookup dictionary,
which caused random unintended effects.

Fix this by implementing proper housekeeping.

Reported-by: Manuel Donaubauer <Manuel.Donaubauer@Uni-Passau.De>
Signed-off-by: Ralf Ramsauer <ralf.ramsauer@oth-regensburg.de>