Mbox: make all emails timezone aware
we need timezone aware datetimes due to the fact, that most of all
emails contain timezone aware timestamps. There's an issue with timezone
unaware timestamps: they can't be compared to timezone aware
timestamps. To cope with that, we simple shift those mails to UTC (which
is also true in most cases).

E.g. python converts this timestamp to an timezone unaware one, while it
is GMT:
  'Fri, 23 Feb 2007 13:35:50 -0000 (GMT)'

Signed-off-by: Ralf Ramsauer <ralf.ramsauer@oth-regensburg.de>