Repository: Improve Diff/Msg parsing
Let MessageDiff only accept lists of strings as input. Furthermore,
remove parse_diff function by Diff constructor.

Also store raw diff message inside the diff, and track the length of the
footer. Unparseable lines at the end of a diff. In case of repositories,
this is an empty newline, in case of mailboxes, this might be the
mailing list footer.

Signed-off-by: Ralf Ramsauer <ralf.ramsauer@oth-regensburg.de>