pasta: new tool: upstream_duration
calculates the upstream duration of patches. Writes results to the file
'upstream_duration' in the following format:

commit_hash days\n
commit_hash days\n
commit_hash days\n
...

where commit_hash is the youngest upstream representative of the class,
and days is the number of days the patch took for being accepted.

Signed-off-by: Ralf Ramsauer <ralf.ramsauer@oth-regensburg.de>