README: improve setup instructions
Pipenv instructions contained a tiny error. But let's use this chance to
replace pipenv by distro-native package installation. Using Ubuntu as a
reference, we only have two packages left that are not in the official
repository: flaks-nav and flask-bootstrap.

At least flask-bootstrap will be removed soon, as it is unmaintained.

Tested with a Ubuntu 18.04 docker container.

Reported-by: Lukas Bulwahn <lukas.bulwahn@gmail.com>
Signed-off-by: Ralf Ramsauer <ralf.ramsauer@oth-regensburg.de>