pasta_analyse: Check reachability of all patches before analysis
An analysis might fail, if a (partial) mbox-result contains emails /
commit hashes that are no longer reacheable with the current
configuration.

Fix this by caching all commits early followed by reporting the number
of patches that could not be cached.

Those patches will then be removed from the result after a timeout of 5
seconds. This gives the user enough time to abort with Ctrl-C.

Reported-by: Lukas Bulwahn <lukas.bulwahn@gmail.com>
Signed-off-by: Ralf Ramsauer <ralf.ramsauer@oth-regensburg.de>