tools: process_mail: Patchwork-IDs don't require brackets
In python, we want to treat them as integers, as they're unique.
Furthermore, we can make use of a trick here: If PATCHWORK_ID isn't set,
it won't hurt us when writing to the file. If it is set, let's put a
whitespace at the beginning of the string.

With this, we can remove the if statement at the end.

Signed-off-by: Ralf Ramsauer <ralf.ramsauer@oth-regensburg.de>