pypasta: mbox: add support to get patchwork id(s) given a message id
Patchwork guarantees that the combination of project and message id is
unique.ie a specific patchwork project will have unique message ids.
Each such message id will be mapped to a single patchwork id.

On Mbox level, multiple patchwork projects might exist. It is possible
for a message id to be part of more than one patchwork project. Thus on
Mbox level a message id will map to a list of patchwork ids.

Signed-off-by: Rohit Sarkar <rohitsarkar5398@gmail.com>