bin: pasta_maintainer_stats: add graph mode
To visualize a graph representing the MAINTAINERS file, we need to first
generate the edge list. This is done by a new mode, graph mode, that
then dumps the resulting edge list to the specified csv file.

We're using networkx to generate the edge list, so it needs to be added
as a new requirement.

Signed-off-by: Pia Eichinger <pia.eichinger@hotmail.de>
[ralf: reword documentation, maintain README.md]
Signed-off-by: Ralf Ramsauer <ralf.ramsauer@oth-regensburg.de>